backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
google-generativeai
bcrypt
pyjwt
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
faiss-cpu
scikit-learn
pypdf
Pillow
numpy
werkzeug
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
resend
APScheduler
